feat(templates): improve Boilerplate concurrency handling #10777 (#10779)

+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Controllers/Identity/RoleManagementController.cs
@@ -76,6 +76,9 @@ public partial class RoleManagementController : AppControllerBase, IRoleManageme
     {
         var role = await GetRoleByIdAsync(roleDto.Id, cancellationToken);
 
+        if (role.ConcurrencyStamp != roleDto.ConcurrencyStamp)
+            throw new ConflictException();
+
         roleDto.Patch(role);
 
         var result = await roleManager.UpdateAsync(role);
@@ -86,11 +89,14 @@ public partial class RoleManagementController : AppControllerBase, IRoleManageme
         return role.Map();
     }
 
-    [HttpPost("{roleId}")]
+    [HttpPost("{roleId}/{concurrencyStamp}")]
     [Authorize(Policy = AuthPolicies.ELEVATED_ACCESS)]
-    public async Task Delete(Guid roleId, CancellationToken cancellationToken)
+    public async Task Delete(Guid roleId, string concurrencyStamp, CancellationToken cancellationToken)
     {
         var role = await GetRoleByIdAsync(roleId, cancellationToken);
+
+        if (role.ConcurrencyStamp != concurrencyStamp)
+            throw new ConflictException();
 
         await roleManager.DeleteAsync(role);
     }
@@ -199,7 +205,7 @@ public partial class RoleManagementController : AppControllerBase, IRoleManageme
     public async Task SendNotification(SendNotificationToRoleDto dto, CancellationToken cancellationToken)
     {
         //#if (signalR == true)
-        var signalRConnectionIds = await DbContext.UserSessions.Where(us => us.SignalRConnectionId != null && 
+        var signalRConnectionIds = await DbContext.UserSessions.Where(us => us.SignalRConnectionId != null &&
                                                                             us.User!.Roles.Any(r => r.RoleId == dto.RoleId))
                                                                .Select(us => us.SignalRConnectionId!).ToArrayAsync(cancellationToken);
